logsumexp
