+++ tfjs-core/src/public/chained_ops/register_all_chained_ops.ts
@@ -47,6 +47,7 @@ import './logical_and';
 import './logical_not';
 import './logical_or';
 import './logical_xor';
+import './log_sum_exp';
 import './mat_mul';
 import './max';
 import './max_pool';
