+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -57,6 +57,7 @@ const CHAINED_OPS = [
   'logicalNot',
   'logicalOr',
   'logicalXor',
+  'logSumExp',
   'matMul',
   'max',
   'maximum',
