+++ tfjs-core/src/tests.ts
@@ -95,6 +95,7 @@ import './ops/less_equal_test';
 import './ops/less_test';
 import './ops/local_response_normalization_test';
 import './ops/log_loss_test';
+import './ops/log_sum_exp_test';
 import './ops/logical_and_test';
 import './ops/logical_not_test';
 import './ops/logical_or_test';
